Modified 'move caret and insert =' quick fix analogous to StatementEffectIntroduceVariableQuickFix.
